refactor PdfRedactionModel usage
- fix modal-on-top-of-modal weirdness,
- including Esc key closing only one modal
